# models/bundler_account.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// BundlerAccount bundler account
//
// swagger:model BundlerAccount
type BundlerAccount struct {

	// The address of the bundler
	Address string `json:"address,omitempty"`
}

// Validate validates this bundler account
func (m *BundlerAccount) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this bundler account based on context it is used
func (m *BundlerAccount) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *BundlerAccount) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *BundlerAccount) UnmarshalBinary(b []byte) error {
	var res BundlerAccount
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}